Many backend changes (NOT WORKING)
- Websocket now authenticates you
- DB has some fallbacks

+++ startup/service/db.js
@@ -21,15 +21,24 @@ const gameCollection = db.collection("game");
 })();
 
 async function getUser(email) {
+  if (!email) {
+    return null;
+  }
   return userCollection.findOne({ email: email });
 }
 
 async function getUserByToken(token) {
+  if (!token) {
+    return null;
+  }
   return userCollection.findOne({ token: token });
 }
 
 // update trophies updates the number of trophies the user has.
 async function addTrophies({ email, trophies } = newTrophyData) {
+  if (!email || !trophies) {
+    return null;
+  }
   await userCollection.updateOne(
     { email: email },
     { $inc: { trophies: trophies } }
@@ -37,24 +46,39 @@ async function addTrophies({ email, trophies } = newTrophyData) {
 }
 
 async function addUser(user) {
+  if (!user) {
+    return null;
+  }
   await userCollection.insertOne(user);
 }
 
 async function updateUser(user) {
+  if (!user) {
+    return null;
+  }
   await userCollection.updateOne({ email: user.email }, { $set: user });
 }
 
 async function getGame(roomCode) {
+  if (!roomCode) {
+    return null;
+  }
   return gameCollection.findOne({ roomCode: roomCode });
 }
 
 function getGameByPlayerEmail(email) {
+  if (!email) {
+    return null;
+  }
   let match = { players: { $elemMatch: { email: email } } };
   // console.log(match);
   return gameCollection.findOne(match);
 }
 
 async function setGame(roomCode, game) {
+  if (!roomCode || !game) {
+    return null;
+  }
   await gameCollection.updateOne(
     { roomCode: roomCode },
     { $set: game },
@@ -63,6 +87,9 @@ async function setGame(roomCode, game) {
 }
 
 async function deleteGame(roomCode) {
+  if (!roomCode) {
+    return null;
+  }
   await gameCollection.deleteOne({ roomCode: roomCode });
 }
 
@@ -76,17 +103,20 @@ async function getAllGames() {
 }
 
 async function pushStory(roomCode, story) {
+  if (!roomCode || !story) {
+    return null;
+  }
   await gameCollection.updateOne(
     { roomCode: roomCode },
     { $push: { stories: story } }
   );
 }
 
-async function getHighScores() {
+async function getHighScores(limit = 10) {
   const query = { trophies: { $gt: 0, $lt: 900 } };
   const options = {
     sort: { trophies: -1 },
-    limit: 10,
+    limit: limit,
   };
   const cursor = userCollection.find(query, options);
   return cursor.toArray();
